LCSAJdump/requirements.txt
capstone>=5.0.0
pyelftools>=0.29
networkx>=3.0
click>=8.0
